Additional namespace determined from caller in Specifaication.__init__.
If a spec needs additional context, i. e. it contains names other than the
name of the candidate or the names of builtins, a corresponding namespace can
be given as parameter. If no namespace is given, it is determined from the
namespace in which the spec is created and then stored in the spec itself. It is
used when the spec is evaluated.

[Fixes: Issue #1 Implicit context from caller results in exception]